+++ linphone-app/src/components/chat-room/ChatRoomModel.hpp
@@ -82,6 +82,7 @@ public:
 	Q_PROPERTY(QString username READ getUsername NOTIFY usernameChanged)
 	Q_PROPERTY(QString avatar READ getAvatar NOTIFY avatarChanged)
 	Q_PROPERTY(int presenceStatus READ getPresenceStatus NOTIFY presenceStatusChanged)
+	Q_PROPERTY(QDateTime presenceTimestamp READ getPresenceTimestamp NOTIFY presenceStatusChanged)
 	Q_PROPERTY(LinphoneEnums::ChatRoomState state READ getState NOTIFY stateChanged)
 	
 	Q_PROPERTY(long ephemeralLifetime READ getEphemeralLifetime WRITE setEphemeralLifetime NOTIFY ephemeralLifetimeChanged)
@@ -120,6 +121,7 @@ public:
 	QString getUsername () const;
 	QString getAvatar () const;
 	int getPresenceStatus() const;
+	QDateTime getPresenceTimestamp() const;
 	LinphoneEnums::ChatRoomState getState() const;
 	bool isReadOnly() const;
 	bool isEphemeralEnabled() const;