+++ linphone-app/src/components/presence/OwnPresenceModel.cpp
@@ -75,7 +75,7 @@ QVariantList OwnPresenceModel::getStatuses () const {
   addBuildStatus(statuses, Presence::Online);
   addBuildStatus(statuses, Presence::Busy);
   addBuildStatus(statuses, Presence::DoNotDisturb);
-  addBuildStatus(statuses, Presence::Offline);
+  //addBuildStatus(statuses, Presence::Offline);	// Do not propose this status as it is not fully supported
 
   return statuses;
 }