+++ linphone-app/src/utils/Constants.cpp
@@ -124,6 +124,7 @@ constexpr char Constants::LinphoneDomain[];
 constexpr char Constants::DefaultContactParameters[];
 constexpr char Constants::DefaultContactParametersOnRemove[];
 constexpr int Constants::DefaultExpires;
+constexpr int Constants::DefaultPublishExpires;
 constexpr char Constants::DownloadUrl[];
 constexpr char Constants::VersionCheckReleaseUrl[];
 constexpr char Constants::VersionCheckNightlyUrl[];